refactor: folder 구조 변경 [AIR-126] - cmd -> command - RoomPhotoService 수정
