+++ aircnc/src/main/java/com/gurudev/aircnc/domain/member/service/MemberService.java
@@ -3,7 +3,7 @@ package com.gurudev.aircnc.domain.member.service;
 import com.gurudev.aircnc.domain.member.entity.Email;
 import com.gurudev.aircnc.domain.member.entity.Member;
 import com.gurudev.aircnc.domain.member.entity.Password;
-import com.gurudev.aircnc.domain.member.service.cmd.MemberCommand.MemberRegisterCommand;
+import com.gurudev.aircnc.domain.member.service.command.MemberCommand.MemberRegisterCommand;
 
 public interface MemberService {
