+++ aircnc/src/main/java/com/gurudev/aircnc/domain/member/service/MemberServiceImpl.java
@@ -4,7 +4,7 @@ import com.gurudev.aircnc.domain.member.entity.Email;
 import com.gurudev.aircnc.domain.member.entity.Member;
 import com.gurudev.aircnc.domain.member.entity.Password;
 import com.gurudev.aircnc.domain.member.repository.MemberRepository;
-import com.gurudev.aircnc.domain.member.service.cmd.MemberCommand.MemberRegisterCommand;
+import com.gurudev.aircnc.domain.member.service.command.MemberCommand.MemberRegisterCommand;
 import com.gurudev.aircnc.exception.AircncRuntimeException;
 import com.gurudev.aircnc.exception.NotFoundException;
 import com.gurudev.aircnc.infrastructure.security.PasswordEncryptor;
