+++ aircnc/src/main/java/com/gurudev/aircnc/domain/room/service/RoomPhotoService.java
@@ -8,4 +8,6 @@ public interface RoomPhotoService {
 
   List<RoomPhoto> upload(List<MultipartFile> multipartFiles);
 
+  RoomPhoto upload(MultipartFile multipartFile);
+
 }
