+++ aircnc/src/main/java/com/gurudev/aircnc/domain/room/service/RoomService.java
@@ -1,7 +1,7 @@
 package com.gurudev.aircnc.domain.room.service;
 
 import com.gurudev.aircnc.domain.room.entity.Room;
-import com.gurudev.aircnc.domain.room.service.cmd.RoomCommand.RoomRegisterCommand;
+import com.gurudev.aircnc.domain.room.service.command.RoomCommand.RoomRegisterCommand;
 import java.util.List;
 
 public interface RoomService {
